Create README.md
This commit introduces a detailed README for my GitHub profile, providing an overview of who I am, my professional background, and the goals I am currently pursuing in the tech industry. It includes a description of my transition from healthcare to tech, my areas of focus such as AI, Cybersecurity, and Data Analysis, and the projects I’m working on. The README is designed to give visitors an insight into my journey, passion, and aspirations as I continue to learn and grow in the software development space.